+++ game-dishonored2/CHANGELOG.md
@@ -5,6 +5,10 @@
 - Better installer for D2Fix mod - handle multiple folders + select options based on game version.
 - ? Disable Config/Save installers so that hardlinks will be available for more users? - symlinks seem fine.
 
+## [0.6.1] - 2025-11-12
+
+- Fixed bug in Save modType registration that could break deployment in other games.
+
 ## [0.6.0] - 2025-11-10
 
 - Improved installer for Void Installer mods.